Move CommunityStateSchema to structs package
