+++ cmd/habitat/community/state/state_test.go
@@ -5,17 +5,18 @@ import (
 	"encoding/json"
 	"testing"
 
+	"github.com/eagraf/habitat/structs/community"
 	"github.com/qri-io/jsonschema"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestCommunityStateSchema(t *testing.T) {
 	rs := &jsonschema.Schema{}
-	err := json.Unmarshal(communityStateSchema, rs)
+	err := json.Unmarshal(community.CommunityStateSchema, rs)
 	assert.Nil(t, err)
 
 	// Test that an empty CommunityState struct matches the schema
-	cs := CommunityState{}
+	cs := community.CommunityState{}
 	marshaled, err := json.Marshal(&cs)
 	assert.Nil(t, err)
 	keyErrs, err := rs.ValidateBytes(context.Background(), marshaled)
@@ -32,7 +33,7 @@ func TestJSONState(t *testing.T) {
 	initState := []byte(`{
 		"community_id": "abc"
 	}`)
-	jsonState, err := NewJSONState(communityStateSchema, initState)
+	jsonState, err := NewJSONState(community.CommunityStateSchema, initState)
 	assert.Nil(t, err)
 
 	patch := []byte(`[{
@@ -46,7 +47,7 @@ func TestJSONState(t *testing.T) {
 }
 
 func TestCounterIncrement(t *testing.T) {
-	jsonState, err := NewJSONState(communityStateSchema, []byte(`{"community_id":"abc"}`))
+	jsonState, err := NewJSONState(community.CommunityStateSchema, []byte(`{"community_id":"abc"}`))
 	assert.Nil(t, err)
 
 	csm := NewCommunityStateMachine(jsonState, &LocalDispatcher{jsonState: jsonState})
